requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.3
cffi==1.16.0
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
email-validator==2.1.0.post1
Faker==23.3.0
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
jaraco.classes==3.3.1
keyring==24.3.0
more-itertools==10.2.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
playwright==1.41.2
pluggy==1.4.0
pycparser==2.21
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
pyee==11.0.1
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
schedule==1.2.1
six==1.16.0
slack_sdk==3.27.1
soupsieve==2.5
toml==0.10.2
tomlkit==0.12.3
typing_extensions==4.10.0
typish==1.9.3
tzdata==2024.1
zipp==3.17.0
